Make changeable query execution target
